separate api from board actions
